#2 extract pod convenience symbols into own file
These structs and methods do not exactly have something to do with
setting up a cluster. So they get their own file
